Add nftOps support to `patchOperationWithHash`
