+++ libs/coin-framework/src/operation.ts
@@ -1,8 +1,16 @@
-import type { Account, AccountLike, Operation } from "@ledgerhq/types-live";
+import invariant from "invariant";
 import { BigNumber } from "bignumber.js";
 import { getEnv } from "@ledgerhq/live-env";
+import type { Account, AccountLike, NFTStandard, Operation } from "@ledgerhq/types-live";
+import { encodeERC1155OperationId, encodeERC721OperationId } from "./nft/nftOperationId";
 import { findSubAccountById, getMainAccount } from "./account/helpers";
-import invariant from "invariant";
+import { decodeAccountId } from "./account";
+import { encodeNftId } from "./nft/nftId";
+
+const nftOperationIdEncoderPerStandard: Record<NFTStandard, (...args: any[]) => string> = {
+  ERC721: encodeERC721OperationId,
+  ERC1155: encodeERC1155OperationId,
+};
 
 export function findOperationInAccount(
   { operations, pendingOperations }: AccountLike,
@@ -96,13 +104,33 @@ export function patchOperationWithHash(operation: Operation, hash: string): Oper
     ...operation,
     hash,
     id: encodeOperationId(operation.accountId, hash, operation.type),
-    subOperations:
-      operation.subOperations &&
-      operation.subOperations.map(op => ({
-        ...op,
-        hash,
-        id: encodeOperationId(op.accountId, hash, op.type),
-      })),
+    subOperations: operation.subOperations
+      ? operation.subOperations.map(op => ({
+          ...op,
+          hash,
+          id: encodeOperationId(op.accountId, hash, op.type),
+        }))
+      : [],
+    nftOperations: operation.nftOperations
+      ? operation.nftOperations.map((nftOp, i) => {
+          const { currencyId } = decodeAccountId(operation.accountId);
+          const nftId = encodeNftId(
+            operation.accountId,
+            nftOp.contract || "",
+            nftOp.tokenId || "",
+            currencyId,
+          );
+          const nftOperationIdEncoder =
+            nftOperationIdEncoderPerStandard[(nftOp?.standard as NFTStandard) || ""] ||
+            nftOperationIdEncoderPerStandard.ERC721;
+
+          return {
+            ...nftOp,
+            hash,
+            id: nftOperationIdEncoder(nftId, hash, nftOp.type, 0, i),
+          };
+        })
+      : [],
   };
 }
